# openapi.yaml
openapi: 3.1.0
info:
  title: Eka Legal Consultancy API
  description: |
    Backend API for Eka Legal Consultancy practice management system.
    Supports client portal, booking system, case management, messaging, and document handling.
  version: 1.0.0
  contact:
    name: Eka Legal Support
    email: support@eka-legal.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.eka-legal.com/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          description: MIME type
        size:
          type: integer
          description: File size in bytes
        uploadedAt:
          type: string
          format: date-time
        uploadedBy:
          type: string
        url:
          type: string
          format: uri
      required:
        - id
        - name
        - type
        - size
        - uploadedAt
        - uploadedBy
        - url
...
    DocumentUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            url:
              type: string
              format: uri

    # ============================================
    # Messaging
    # ============================================
    Message:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderRole:
          type: string
          enum: [client, lawyer]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        read:
          type: boolean
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Document'
      required:
        - id
        - senderId
        - senderName
        - senderRole
        - content
        - timestamp
        - read
    Conversation:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        caseTitle:
          type: string
        participants:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              role:
                type: string
        lastMessage:
          type: string
        lastMessageAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
      required:
        - id
        - caseId
        - caseTitle
        - participants
        - lastMessage
        - lastMessageAt
        - unreadCount
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    # ============================================
    # Notifications
    # ============================================
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [appointment, message, document, case, system]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
        link:
          type: string
      required:
        - id
        - type
        - title
        - message
        - read
        - createdAt

    # ============================================
    # Dashboard
    # ============================================
...
    IntakeFormData:
      type: object
      properties:
        personalInfo:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
            preferredContact:
              type: string
          required:
            - name
            - email
            - phone
            - preferredContact
        caseType:
          type: string
        urgency:
          type: string
        description:
          type: string
        additionalInfo:
          type: object
          properties:
            desiredOutcome:
              type: string
            priorCounsel:
              type: string
        consent:
          type: boolean
      required:
        - personalInfo
        - caseType
        - urgency
        - description
        - consent
    IntakeFormResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            caseId:
              type: string
        message:
          type: string
paths:
...
  /cases/{caseId}/documents:
    get:
      tags:
        - Documents
      summary: Get case documents
      description: Retrieve all documents for a specific case
      operationId: getDocumentsByCase
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Documents retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Document'

    post:
      tags:
        - Documents
      summary: Upload document
      description: Upload a document to a case
      operationId: uploadDocument
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Invalid file
  /documents/{documentId}:
    delete:
      tags:
        - Documents
      summary: Delete document
      description: Delete a document by ID
      operationId: deleteDocument
      security:
        - BearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Document not found

  # ============================================
  # Message Endpoints
  # ============================================
  /messages/conversations:
    get:
      tags:
        - Messages
      summary: Get conversations
      description: Retrieve all conversations for the authenticated user
      operationId: getConversations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Conversations retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Conversation'
  /messages/conversations/{conversationId}/messages:
    get:
      tags:
        - Messages
      summary: Get messages
      description: Retrieve messages for a specific conversation
      operationId: getMessages
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'

    post:
      tags:
        - Messages
      summary: Send message
      description: Send a message in a conversation
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - content
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
  /messages/read:
    post:
      tags:
        - Messages
      summary: Mark messages as read
      description: Mark multiple messages as read
      operationId: markMessagesAsRead
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messageIds:
                  type: array
                  items:
                    type: string
              required:
                - messageIds
      responses:
        '200':
          description: Messages marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================
  # Notification Endpoints
  # ============================================
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      description: Retrieve all notifications for the authenticated user
      operationId: getNotifications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notifications retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
  /notifications/{notificationId}/read:
    post:
      tags:
        - Notifications
      summary: Mark notification as read
      description: Mark a specific notification as read
      operationId: markNotificationAsRead
      security:
        - BearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/read-all:
    post:
      tags:
        - Notifications
      summary: Mark all as read
      description: Mark all notifications as read
      operationId: markAllNotificationsAsRead
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================
  # Dashboard Endpoints
  # ============================================
...
  /intake:
    post:
      tags:
        - Intake
      summary: Submit intake form
      description: Submit a client intake form to request legal consultation
      operationId: submitIntakeForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntakeFormData'
      responses:
        '201':
          description: Intake form submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntakeFormResponse'
        '400':
          description: Invalid form data
  /intake/draft:
    get:
      tags:
        - Intake
      summary: Get intake draft
      description: Retrieve saved intake form draft
      operationId: getIntakeDraft
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Draft retrieved (null if no draft exists)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/IntakeFormData'
                          - type: 'null'

    post:
      tags:
        - Intake
      summary: Save intake draft
      description: Save intake form progress as a draft
      operationId: saveIntakeDraft
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntakeFormData'
      responses:
        '200':
          description: Draft saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'